HarmonyEcoSystem/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
SQLAlchemy==2.0.20
PyYAML==6.0.1
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
zeep==4.2.1
lxml==4.9.3
requests==2.31.0
openpyxl==3.1.2
Pillow==10.3.0
